views list with core fields metadata (backend viewId)

+++ src/app/_modules/list/_components/list-datatable/list-datatable.component.ts
@@ -1,16 +1,16 @@
 import { SelectionModel } from '@angular/cdk/collections';
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
-import { ListPageViewDetails } from '@models/list-page/listpage';
+import { ListPageViewDetails, ViewsPage } from '@models/list-page/listpage';
 import { Userdetails } from '@models/userdetails';
 import { SharedServiceService } from '@modules/shared/_services/shared-service.service';
 import { ListService } from '@services/list/list.service';
-import { SchemaDetailsService } from '@services/home/schema/schema-details.service';
 import { UserService } from '@services/user/userservice.service';
 import { BehaviorSubject, Subscription } from 'rxjs';
-import { MetadataModeleResponse } from '@models/schema/schemadetailstable';
 import { PageEvent } from '@angular/material/paginator';
 import { ListDataSource } from './list-data-source';
+import { FieldMetaData } from '@models/core/coreModel';
+import { CoreService } from '@services/core/core.service';
 
 
 @Component({
@@ -31,10 +31,9 @@ export class ListDatatableComponent implements OnInit, OnDestroy {
   currentView: ListPageViewDetails = new ListPageViewDetails();
 
   /**
-   * Default view to display
+   * Default view to display if there is no configured view
    */
   defaultView = {
-    viewId: 'default',
     viewName: 'Default view',
     isSystemView: false,
     fieldsReqList: [
@@ -48,7 +47,7 @@ export class ListDatatableComponent implements OnInit, OnDestroy {
   /**
    * hold views list
    */
-  viewsList: ListPageViewDetails[] = [];
+  viewsList: ViewsPage = new ViewsPage();
 
   /**
    * for views list paging
@@ -77,7 +76,7 @@ export class ListDatatableComponent implements OnInit, OnDestroy {
 
   userDetails = new Userdetails();
 
-  metadataFldLst: MetadataModeleResponse;
+  metadataFldLst: FieldMetaData[] = [];
 
 
   constructor(
@@ -86,7 +85,7 @@ export class ListDatatableComponent implements OnInit, OnDestroy {
     private sharedServices: SharedServiceService,
     private userService: UserService,
     private listService: ListService,
-    private schemaDetailsService: SchemaDetailsService) {
+    private coreService: CoreService) {
 
     this.dataSource = new ListDataSource(this.listService);
 
@@ -110,16 +109,12 @@ export class ListDatatableComponent implements OnInit, OnDestroy {
       this.getFldMetadata();
     });
 
-    this.sharedServices.getViewDetailsData().subscribe(view => {
-      if (view && view.viewId) {
-        const index = this.viewsList.findIndex(v => v.viewId === view.viewId);
-        if (index !== -1) {
-          this.viewsList[index] = view;
-        } else {
-          this.viewsList.push(view);
-        }
-
-        this.updateTableColumns(view.viewId);
+    this.sharedServices.getViewDetailsData().subscribe(resp => {
+      if (resp && resp.isUpdate) {
+        this.currentView = resp.viewDetails;
+        this.updateTableColumns();
+      } else {
+        this.getViewsList();
       }
     })
 
@@ -135,29 +130,44 @@ export class ListDatatableComponent implements OnInit, OnDestroy {
     } else {
       this.viewsPageIndex = 0;
     }
-    const sub = this.listService.getAllListPageViews(this.userDetails.userName, this.userDetails.currentRoleId, this.userDetails.plantCode, this.moduleId, this.viewsPageIndex, 5)
+    const sub = this.listService.getAllListPageViews(this.userDetails.userName, this.userDetails.currentRoleId, this.userDetails.plantCode, this.moduleId, this.viewsPageIndex)
       .subscribe(views => {
 
-        if (views && views.length) {
+        if(views && ((views.userViews && views.userViews.length) || (views.systemViews && views.systemViews.length))) {
           if (loadMore) {
-            this.viewsList = this.viewsList.concat(views);
+            this.viewsList.userViews = this.viewsList.userViews.concat(views.userViews || []);
+            this.viewsList.systemViews = this.viewsList.systemViews.concat(views.systemViews || []);
           } else {
             this.viewsList = views;
-            this.updateTableColumns(this.viewsList[0].viewId);
+            const defaultViewId = this.getDefaultViewId();
+            this.getViewDetails(defaultViewId);
           }
         } else if (!loadMore) {
-            this.viewsList = [this.defaultView];
-            this.updateTableColumns(this.viewsList[0].viewId);
+            this.currentView = this.defaultView;
+            this.updateTableColumns();
         }
-
       }, error => {
         console.error(`Error :: ${error.message}`);
         if(!loadMore) {
-          this.viewsList = [this.defaultView];
-          this.updateTableColumns(this.viewsList[0].viewId);
+          this.currentView = this.defaultView;
+          this.updateTableColumns();
         }
       });
 
+    this.subscriptionsList.push(sub);
+  }
+
+  /**
+   * get view details by id
+   */
+   getViewDetails(viewId) {
+    const sub =  this.listService.getListPageViewDetails(viewId, this.userDetails.userName, this.userDetails.currentRoleId, this.userDetails.plantCode, this.moduleId)
+      .subscribe(response => {
+          this.currentView = response;
+          this.updateTableColumns();
+    }, error => {
+      console.error(`Error : ${error.message}`);
+    });
     this.subscriptionsList.push(sub);
   }
 
@@ -166,6 +176,10 @@ export class ListDatatableComponent implements OnInit, OnDestroy {
    * @param viewId to delete
    */
   deleteView(viewId: string) {
+
+    if(!viewId || viewId === 'default' ) {
+      return;
+    }
     const sub = this.listService.deleteListPageView(viewId, this.userDetails.userName, this.userDetails.currentRoleId, this.userDetails.plantCode, this.moduleId)
       .subscribe(resp => {
         this.getViewsList();
@@ -183,7 +197,7 @@ export class ListDatatableComponent implements OnInit, OnDestroy {
     if (this.moduleId === undefined || this.moduleId.trim() === '') {
       throw new Error('Module id cant be null or empty');
     }
-    const sub = this.schemaDetailsService.getMetadataFields(this.moduleId).subscribe(response => {
+    const sub = this.coreService.getAllFieldsForView(this.moduleId).subscribe(response => {
       this.metadataFldLst = response;
     }, error => {
       console.error(`Error : ${error.message}`);
@@ -229,22 +243,20 @@ export class ListDatatableComponent implements OnInit, OnDestroy {
    * open view config sidesheet
    */
   openTableViewSettings(edit?: boolean) {
-    const viewId = edit && this.currentView.viewId &&  this.currentView.viewId !== 'default' && !this.currentView.viewId.includes('view') ?  this.currentView.viewId : 'new';
+    const viewId = edit && this.currentView.viewId ?  this.currentView.viewId : 'new';
     this.router.navigate([{ outlets: { sb: `sb/list/table-view-settings/${this.moduleId}/${viewId}` } }], { queryParamsHandling: 'preserve' });
   }
 
   /**
    * update displayed columns on view change
-   * @param viewId selected view id
    */
-  updateTableColumns(selectedViewId: string) {
-    const activeView = this.viewsList.find(v => v.viewId === selectedViewId);
-    if (activeView && activeView.fieldsReqList) {
+  updateTableColumns() {
 
-      this.currentView = activeView;
+    if (this.currentView && this.currentView.fieldsReqList) {
+
       this.getTableData();
 
-      const activeColumns: string[] = activeView.fieldsReqList.map(field => field.fieldId);
+      const activeColumns: string[] = this.currentView.fieldsReqList.map(field => field.fieldId);
       this.displayedColumns.next(this.staticColumns.concat(activeColumns));
     }
 
@@ -254,7 +266,7 @@ export class ListDatatableComponent implements OnInit, OnDestroy {
    * get table data records
    */
   getTableData() {
-    const viewId = this.currentView.viewId &&  this.currentView.viewId !== 'default' && !this.currentView.viewId.includes('view') ?  this.currentView.viewId : '';
+    const viewId = this.currentView.viewId ?  this.currentView.viewId : '';
     this.dataSource.getData(this.moduleId, viewId, this.recordsPageIndex);
   }
 
@@ -271,14 +283,32 @@ export class ListDatatableComponent implements OnInit, OnDestroy {
    * get system views list
    */
   get systemViews() {
-    return this.viewsList.filter(view => view.isSystemView);
+    return this.viewsList.systemViews || [];
   }
 
   /**
    * get user views list
    */
   get userViews() {
-    return this.viewsList.filter(view => !view.isSystemView);
+    return this.viewsList.userViews || [];
+  }
+
+  /**
+   * get user's default view id
+   */
+  getDefaultViewId(): string {
+    const view = this.userViews.concat(this.systemViews).find(v => v.default);
+    return view ? view.viewId : this.userViews.concat(this.systemViews)[0].viewId;
+  }
+
+  /**
+   * get field description based on field id
+   * @param fieldId field id
+   * @returns field description
+   */
+  getFieldDesc(fieldId: string): string {
+    const field = this.metadataFldLst.find(f => f.fieldId === fieldId);
+    return field ? field.fieldDescri || 'Unkown': fieldId || 'Unkown';
   }
 
   /**
